requirements.txt
streamlit
langchain
tiktoken
pinecone-client
pypdf
google-generativeai
langchain-google-genai
python-dotenv
unstructured